Remove some unnecessary objects around metrics handling for simplicity.
