--- admin/app/conf/ExecutionContexts.scala
@@ -1,5 +1,0 @@
-package conf
-
-trait ExecutionContexts {
-  implicit lazy val executionContext = play.api.libs.concurrent.Execution.Implicits.defaultContext
-}
