+++ article/app/conf/context.scala
@@ -1,20 +1,16 @@
 package conf
 
-import common._
-import _root_.play.api.{ Application => PlayApp }
-import com.gu.management.play._
-import com.gu.management._
-import logback.LogbackLevelPage
-import contentapi.ContentApiMetrics
+import common.Metrics
+import com.gu.management.{ PropertiesPage, StatusPage, ManifestPage }
+import com.gu.management.play.{ Management => GuManagement }
+import com.gu.management.logback.LogbackLevelPage
+import play.api.{ Application => PlayApp }
 
 class SwitchBoardPlugin(app: PlayApp) extends SwitchBoardAgent(Configuration, Switches.all)
 
-object Metrics {
-  val all: Seq[Metric] = ContentApiMetrics.all ++ CommonMetrics.all
-}
-
-object Management extends Management {
+object Management extends GuManagement {
   val applicationName = "frontend-article"
+  val metrics = Metrics.contentApi ++ Metrics.common
 
   lazy val pages = List(
     new ManifestPage,
@@ -22,7 +18,7 @@ object Management extends Management {
       "/world/2012/sep/11/barcelona-march-catalan-independence",
       "/sport/2012/nov/17/becky-james-jess-varnish-rio-olympics"
     ),
-    StatusPage(applicationName, Metrics.all),
+    StatusPage(applicationName, metrics),
     new PropertiesPage(Configuration.toString),
     new LogbackLevelPage(applicationName)
   )
