+++ front/app/conf/context.scala
@@ -1,25 +1,21 @@
 package conf
 
+import common.Metrics
+import com.gu.management.{ PropertiesPage, StatusPage, ManifestPage }
+import com.gu.management.play.{ Management => GuManagement }
+import com.gu.management.logback.LogbackLevelPage
 import play.api.{ Application => PlayApp }
-import com.gu.management._
-import com.gu.management.play._
-import logback.LogbackLevelPage
-import common._
-import contentapi.ContentApiMetrics
 
 class SwitchBoardPlugin(app: PlayApp) extends SwitchBoardAgent(Configuration, Switches.all)
 
-object Metrics {
-  val all: Seq[Metric] = ContentApiMetrics.all ++ CommonMetrics.all
-}
-
-object Management extends Management {
+object Management extends GuManagement {
   val applicationName = "frontend-front"
+  val metrics = Metrics.contentApi ++ Metrics.common
 
   lazy val pages = List(
     new ManifestPage,
     new UrlPagesHealthcheckManagementPage("/"),
-    StatusPage(applicationName, Metrics.all),
+    StatusPage(applicationName, metrics),
     new PropertiesPage(Configuration.toString),
     new LogbackLevelPage(applicationName)
   )
